refactor(components): simplify section component
The previous implementation was better as it didn't rely on a hardcoded
background color for the section title. However, it had off-by-one
pixels depending on the browser's zoom level.

This commit's approach hardcodes the title's background color to make it
seem like the gaps in the top border are there. This is good enough for
my site since I won't be using gradient or image as the background.