refactor(utils): add utility typewriter class
Unfortunately, Svelte's transitions are not very flexible. For example,
with key blocks, the new element is mounted without waiting for the
out transition of the previous element to end, so both elements will
be visible at the same time for a period. They are also not
interruptible.

This commit adds a new `Typewriter` class that implements the same
typewriter animation. This implementation uses a text buffer to
contain the state of the animated text. The text buffer is reactive, so
callers can display it to show the current state of the animation. The
implementation also has support for interrupting the animation and smarter
deletion and typing.