fix: fix page content container having 0 height
Using a grid seems to be a better approach than using absolutely
positioned elements for each page's content. Using `absolute` means
the element can't influence the layout of the rest of the page, which
causes weird margin issues (the root content container has 0 height)
when the content overflows. Using a grid allows them to still overlap
and work around the transition DOM issue, while allowing them to be part
of the page's layout.