fix(contact): fix abrupt movement of contents when transitioning from projects page
Because each page's content container can now affect the layout of
the page, the project cards on the projects page cause the shared root
container's height to grow. Since the vertical space above the contact cards
is based on the container's height, when transitioning from the projects page to
the contact page, the height will abruptly change once the projects page's container
is removed from the DOM, which will then cause the vertical space to
suddenly change in height.

This commit avoids the whole flexbox shenanigans and instead bases the
vertical space on vh units.