refactor: use more robust approach to preventing self-navigation
Instead of trying to hack around Astro's ClientRouter's logic, it's
easier and more reliable to just override the handler for links.

While the previous approach solves the self-navigation issue for the
majority of cases, there were still a few edge cases that were buggy.
For example, if you spam click on different links, you can end up with a
mismatch between the URL shown in the browser and the actual content
displayed. Avoiding the client router's logic altogether by overriding
the click handler for the links seems to prevent these issues.